refactor(hooks): hooks/onchange をフォルダへ整理する
hooks.rs + onchange.rs（フラット2ファイル）を、apply.rs+apply/・
locals.rs+locals/ と同じ「薄いオーケストレータ umbrella + 役割別
children」の慣習に合わせて src/core/hooks/ フォルダへ整理する。

- hooks.rs（umbrella）: 頻度ディスパッチ（run_unit_hooks / run_onchange /
  run_always）
- hooks/exec.rs（新規）: プロセス実行プリミティブ（exec /
  is_bare_command / program_path）
- hooks/onchange.rs（移動）: 状態永続化＋指紋（State / hash_dir /
  short_hash）

Hook/Frequency は Kind/Strategy と同じ理由（スキーマ型は manifest.rs に
一元化）で manifest.rs に残置。ロジックは無変更（純粋な構造リファクタ）。

あわせて hooks/exec.rs の契約テストが実ツール名 "bat" を fixture に
使っていた箇所を、D9（設計書§15: 契約テストは架空 fixture のみを使う）
に従い "foo" へ置き換えた。
