+++ src/core/discover.rs
@@ -62,7 +62,7 @@ fn walk(dir: &Path, source: &Path, out: &mut Vec<Unit>) -> Result<(), String> {
 ///
 /// §6.3 の委譲規則を 1 か所に集約する: `manifest.toml` 自体と、別の `manifest.toml` を持つ
 /// サブツリー（委譲先の別単位）は除外する（[`crate::core::apply::copy`] の `copy_tree` と同じ範囲）。onchange
-/// ハッシュ（[`crate::core::onchange::hash_dir`]）が「この単位のソースが変わったか」を判定する材料に使う。
+/// ハッシュ（[`crate::core::hooks::onchange::hash_dir`]）が「この単位のソースが変わったか」を判定する材料に使う。
 pub fn unit_files(dir: &Path) -> Result<Vec<PathBuf>, String> {
     let mut files = Vec::new();
     collect_unit_files(dir, &mut files)?;
